extract reproducible wheel stuff
